Overwrite clickable grid with console version
Needed to take a step back and approach the tic tac toe problem
from a different angle. Focusing on creating a console version
of the game before adding UI. Winning/tie logic not in place,
but the GameController initializes the game, players can add
mark on grid, and GameController switches players.